Big refactoring of how METH are defined and declared
The goal is make easier and less error prone the experience of porting
existing C-API modules to HPy.

So by default, the existing METH_NOARGS, METH_O, etc. are used for LEGACY
functions which are still implemented with old calling convenction. To use the
new HPy calling convention, you must explicitly declare methods as HPy_METH_*.

This should make it possible to port modules to HPy one function at a time,
without having to touch everywhere just to be able to compile.

Because of this, we had to rename some of the existing macros, so the new
convention is:

  - HPy_METH_NOARGS: this is the flag to use inside HPyMethodDef

  - HPy_DEF_METH_NOARGS(name): this is the macro to use to generate the
    trampolines

Eventually, we will probably introduce also this:

  - HPy_DECL_METH_NOARGS: this will *declare* the trampolines, so they can be
    referenced across multiple C files, which is not supported using the
    current approach because we declare everything as static.