WIP: start yet another refactoring, hopefully the last of this kind.
After an IRC conversation with @arigo, we decided to generalize the concept of
HPyMeth/HPySlot: modules and types have an array of HPyDef[], which can be of
many kinds: for now, we implemented only HPyMeth, which are the equivalent of
PyMethodDef[].

In the next commit, we will fix/implement slots, which contain all the "other"
slots which are not PyMethodDef, PyMemberDef and PyGetSetDef.

Eventually, we will implement also the latter two, and possibly other kinds if
the need arises.

The end result is that modules and types contain a "flat" list of definitions:
from the API point of view this is better than the old API, because the
separation between PyMethoDef/PyMemberDef/PyGetSetDef/all other slots is
ultimately a CPython implementation detail.