(antocuni, arigo, ronan) This is probably going to be the HPy commit which was done at the highest altitude: 3317m, (Teodulo refuge, Italy).
Reintroduce HPyObject_HEAD: this makes it MUCH easier for people to port
existing code, and we can remove the has_pyobject_head hack, because it was
impossible to implement HPy_Cast efficiently in presence of it.

Alternate implementations like PyPy can avoid wasting the extra space as
described by the comment inside the code.