Implement hpy.universal.load(..., debug=True)
It is now possible to load an universal module in debug mode: in that case
hpy.universal imports hpy.debug._ctx and asks it to wrap the global
context. At the moment the only difference is that the debug ctx has a
different ctx->name, but this is enough to make test_debug.test_ctx_name
passing!

Moreover, the hpy_abi pytest fixture now can 'cpython', 'universal' or
'debug', which means that all tests will be automatically tested also in debug
mode.  In theory, we could cache the result of the compilation since
'universal' and 'debug' compiles the very same .hpy.so file, but on my machine
the compilation itself seems very fast, so I'm not sure it is worth the
complexity. We might want to revisit it in the future.