Completely change the approach and greatly simplify the code.
The original approach was to have a separate hpy.debug module which wraps a
generic ctx into a debug ctx; hpy.debug could in theory be distributed
independently than hpy.universal and be used both by CPython, PyPy,
GraalPython, etc.

The theory was nice but in practice it made the code and file structure
unnecessarily complex. The new approach is:

- keep hpy/debuc/src/*.c: this is still a generic implementation of the debug
  mode which wraps a generic context

- bundle the debug code inside hpy.universal

- PyPy/GraalPython/etc. will be able to vendor hpy/debug/src/*.c and compile
  it together with their own implementation of HPy