WIP: refactor everything again.
The idea of making DHPy and HPy two different C types was nice in theory, but
it was too complicated and confusing in practice. The new approach of using
DHPy and UHPy is less type safe but probably easier to follow. See the comment
inside debug_internal.h for more details