Detect the usage of closed handles.
Ideally, we would like to raise a clean exception when this happens, but it is
impossible to do in a general way because some of the API functions can't
fail. I tried to raise an exception in some cases and abort in others, but
it's too complicate and probably not so useful: using a closed handle is more
or less equivalent to dereference a NULL pointer, so a clean abort is probably
a reasonable thing to do anyway.

However, I also added the possibility of adding a custom hook, and in that
case the execution does NOT abort. This is useful:

1. in tests

2. on CPython, when porting existing modules to HPy. On CPython "closed
   handles" are still valid in practice as long as the refcount > 0: it should
   make it easier to port extensions to HPy, e.g. by printing a warning inside
   the callback and let the execution to continue, so that people can fix the
   warnings one by one.