Start to kill implementation.h, step 1.
The current way to implement things like HPy_SetItem_i&co. is overly
complicated: we try to reuse the implementation on Universal and CPython by
using the ugly _HPy_IMPL_NAME and _HPy_IMPL_NAME_NOPREFIX macros, but there is
a simpler way, which is already used for many other functions: write them
directly as ctx_* and manually write the HPy_* wrapper for the CPython ABI
case.

Eventually, the HPy_* wrappers should be autogenerated, though.