Kill HPyAPI_STORAGE and unify the CPython and Universal cases.
Having two different definitions of HPyAPI_STORAGE was useful when
implementation.h was shared between hpy/universal and hpy/cpython
but now that they are separate it is confusing to have a single macro for two
different things.

It's much easiery to clearly separate the roles of ctx_* functions and HPy_*
ones.