Introduce HPy_TPFLAGS_HAVE_GC.
The original idea was to deduce it automatically based on the presence of
tp_traverse, but this is probably an unacceptable overhead on CPython, because
GC-tracked objects are bigger than non-GC objects. So, we must give the user a
way to choose whether it's needed or not, and the easiest/cleanest way is to
mimimc the CPython's Py_TPFLAGS_HAVE_GC.

As an additional nicety to the user, we check that if the flag is set,
tp_traverse must be provided.