Factor out code for running subprocess tests using the correct ABI
Add helper functions to compile module without loading it and to load that
module in a subprocess and run given code snippet.

Add fatal_exit_code pytest fixture to make it reusable in other tests.

Add option '--subprocess-v' for tests to enable logging of out/err of
subprocesses.

Refactor HPy_FatalError tests to use the new infra and add a test that Python
crashes in debug mode on invalid handle when no debug hook is installed