Check leaks/invalid uses of const char ptrs returned from AsUTF8AndSize
Adds two fields to DebugHandle struct to keep any raw data associated with
that handle and the size of the data. When the handle is closed, the
associated data memory region is mprotected from reading/writing.

AsUTF8AndSize returns a copy of the original result and saves that to
the DebugHandle. The copy is also mprotected from reading.