Add support for metaclass in HPyType_FromSpec
This adds a new member of HPyType_SpecParam_Kind and its handling in
HPyType_FromSpec. CPython does not support metaclass in
PyType_FromSpecWithBases and this PR relies on patched CPython version that
provides PyType_FromSpecWithBasesAndMeta, which is very simple extension
of PyType_FromSpecWithBases (diff ~10 lines), but adds a new API.

Doing all the work on HPy side would require pulling PyType_FromSpecWithBases
code to HPy including all the static helper methods. Not only this would
be a large amount of code, but some of the helper methods use internal
APIs.