define HPyListBuilder_IsNull and HPyTupleBuilder_IsNull
these are relevant to check e.g. if the allocation of the underlying list/tuple
have failed. while the HPy*Builder_Set functions guard for that case, the
calculation of their values may imply work and/or side-effects that a user of
the API may not want to trigger if the list is not being build in any case.