add subset of structseq api
This uses a builder similar to tuple and list builders. Struct sequences in
CPython are tuple subclasses that can hide some of their fields from Python
code. This is difficult to represent nicely in universal mode, so instead we use
the namedtuple type which was structsequences were modeled towards for a long
time in CPython anyway. All functions are just helpers around existing HPy API,
so the structseq.{c,h} files are not part of the HPy ABI, only the API.