We no longer need to monky-patch build.has_ext_modules.
All the logic was to replace has_ext_modules with our own version, we our
version is the very same code as the distutils' one.  I'm not even sure
whether it was ever necessary, because distutils does not seem to have changed
since py3.7 at least.

So, I think that this is safe to remove, but I hope it's not a Chesterton's
fence.